+++ android/app/src/debug/java/com/hedvig/app/DevelopmentActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.widget.Button
-import com.hedvig.app.feature.chat.native.NativeChatActivity
+import com.hedvig.app.feature.chat.NativeChatActivity
 import com.hedvig.android.owldroid.graphql.WhatsNewQuery
 import com.hedvig.app.feature.referrals.ReferralsReceiverActivity
 import com.hedvig.app.feature.referrals.ReferralsSuccessfulInviteActivity
