+++ android/app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         tools:replace="android:label">
 
         <activity
-            android:name=".feature.chat.native.NativeChatActivity"
+            android:name=".feature.chat.NativeChatActivity"
             android:theme="@style/Theme.Exponent.Light" />
 
         <activity
@@ -112,24 +112,9 @@
             android:theme="@style/Theme.Exponent.Light"
             android:windowSoftInputMode="adjustResize" />
         <activity
-            android:name=".feature.offer.OfferActivity"
-            android:configChanges="keyboard|keyboardHidden|orientation|screenSize"
-            android:launchMode="singleTask"
-            android:screenOrientation="portrait"
-            android:theme="@style/Theme.Exponent.Light"
-            android:windowSoftInputMode="adjustResize" />
-        <activity
             android:name=".feature.offer.NativeOfferActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenSize"
             android:launchMode="singleTask"
-            android:screenOrientation="portrait"
-            android:theme="@style/Theme.Exponent.Light"
-            android:windowSoftInputMode="adjustResize" />
-        <activity
-            android:name=".feature.chat.ChatActivity"
-            android:configChanges="keyboard|keyboardHidden|orientation|screenSize"
-            android:launchMode="singleTask"
-            android:parentActivityName=".SplashActivity"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.Exponent.Light"
             android:windowSoftInputMode="adjustResize" />
