+++ android/app/src/main/java/com/hedvig/app/feature/chat/AudioRecorderView.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.feature.chat.native
+package com.hedvig.app.feature.chat
 
 import android.Manifest
 import android.content.Context
@@ -9,7 +9,6 @@ import android.util.AttributeSet
 import android.widget.LinearLayout
 import android.widget.TextView
 import com.hedvig.app.R
-import com.hedvig.app.util.extensions.askForPermissions
 import com.hedvig.app.util.extensions.hasPermissions
 import com.hedvig.app.util.extensions.view.remove
 import com.hedvig.app.util.extensions.view.setHapticClickListener
