--- android/app/src/main/java/com/hedvig/app/feature/chat/dto/UploadData.kt
@@ -1,5 +1,0 @@
-package com.hedvig.app.feature.chat.dto
-
-data class UploadData(
-    val uploadFile: UploadFile
-)
