--- android/app/src/main/java/com/hedvig/app/feature/chat/dto/UploadError.kt
@@ -1,5 +1,0 @@
-package com.hedvig.app.feature.chat.dto
-
-data class UploadError(
-    val message: String?
-)
