--- android/app/src/main/java/com/hedvig/app/feature/chat/dto/UploadFile.kt
@@ -1,6 +1,0 @@
-package com.hedvig.app.feature.chat.dto
-
-data class UploadFile(
-    val signedUrl: String?,
-    val key: String?
-)
