--- android/app/src/main/java/com/hedvig/app/feature/chat/dto/UploadResponse.kt
@@ -1,6 +1,0 @@
-package com.hedvig.app.feature.chat.dto
-
-data class UploadResponse(
-    val data: UploadData?,
-    val errors: List<UploadError>?
-)
