+++ android/app/src/main/java/com/hedvig/app/feature/claims/ui/ClaimsViewModel.kt
@@ -3,7 +3,7 @@ package com.hedvig.app.feature.claims.ui
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
 import com.hedvig.android.owldroid.graphql.CommonClaimQuery
-import com.hedvig.app.data.chat.ChatRepository
+import com.hedvig.app.feature.chat.ChatRepository
 import com.hedvig.app.feature.claims.data.ClaimsRepository
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
