+++ android/app/src/main/java/com/hedvig/app/feature/claims/ui/pledge/HonestyPledgeBottomSheet.kt
@@ -1,17 +1,12 @@
 package com.hedvig.app.feature.claims.ui.pledge
 
 import android.app.Dialog
-import android.content.Intent
 import android.os.Bundle
-import android.support.annotation.IdRes
 import android.view.LayoutInflater
-import androidx.navigation.findNavController
 import com.hedvig.app.R
-import com.hedvig.app.feature.chat.ChatActivity
 import com.hedvig.app.feature.claims.service.ClaimsTracker
 import com.hedvig.app.feature.claims.ui.ClaimsViewModel
 import com.hedvig.app.ui.fragment.RoundedBottomSheetDialogFragment
-import com.hedvig.app.util.extensions.proxyNavigate
 import com.hedvig.app.util.extensions.startClosableChat
 import com.hedvig.app.util.extensions.view.setHapticClickListener
 import kotlinx.android.synthetic.main.bottom_sheet_honesty_pledge.*
