+++ android/app/src/main/java/com/hedvig/app/feature/loggedin/ui/BaseTabViewModel.kt
@@ -2,14 +2,13 @@ package com.hedvig.app.feature.loggedin.ui
 
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
-import com.hedvig.app.data.chat.ChatRepository
+import com.hedvig.app.feature.chat.ChatRepository
 import com.hedvig.app.feature.loggedin.service.TabNotificationService
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.rxkotlin.plusAssign
 import timber.log.Timber
-import javax.inject.Inject
 
-class BaseTabViewModel @Inject constructor(
+class BaseTabViewModel constructor(
     private val chatRepository: ChatRepository,
     private val tabNotificationService: TabNotificationService
 ) : ViewModel() {
