+++ android/app/src/main/java/com/hedvig/app/feature/marketing/ui/MarketingActivity.kt
@@ -12,7 +12,7 @@ import android.widget.ProgressBar
 import com.hedvig.android.owldroid.graphql.MarketingStoriesQuery
 import com.hedvig.app.BaseActivity
 import com.hedvig.app.R
-import com.hedvig.app.feature.chat.ChatActivity
+import com.hedvig.app.feature.chat.NativeChatActivity
 import com.hedvig.app.feature.marketing.service.MarketingTracker
 import com.hedvig.app.util.OnSwipeListener
 import com.hedvig.app.util.SimpleOnSwipeListener
@@ -285,7 +285,7 @@ class MarketingActivity: BaseActivity() {
                 marketingStoriesViewModel.page.value,
                 marketingStoriesViewModel.blurred.value
             )
-            val intent = Intent(this, ChatActivity::class.java)
+            val intent = Intent(this, NativeChatActivity::class.java)
             intent.putExtra("intent", "login")
             intent.putExtra("show_restart", true)
             startActivity(intent)
@@ -296,7 +296,7 @@ class MarketingActivity: BaseActivity() {
                 marketingStoriesViewModel.page.value,
                 marketingStoriesViewModel.blurred.value
             )
-            val intent = Intent(this, ChatActivity::class.java)
+            val intent = Intent(this, NativeChatActivity::class.java)
             intent.putExtra("intent", "onboarding")
             intent.putExtra("show_restart", true)
             startActivity(intent)
