+++ android/app/src/main/java/com/hedvig/app/feature/profile/ui/ProfileViewModel.kt
@@ -5,7 +5,7 @@ import android.arch.lifecycle.ViewModel
 import android.net.Uri
 import com.hedvig.android.owldroid.graphql.ProfileQuery
 import com.hedvig.android.owldroid.graphql.RedeemReferralCodeMutation
-import com.hedvig.app.data.chat.ChatRepository
+import com.hedvig.app.feature.chat.ChatRepository
 import com.hedvig.app.feature.profile.data.ProfileRepository
 import com.hedvig.app.service.Referrals
 import com.hedvig.app.service.RemoteConfig
