+++ android/app/src/main/java/com/hedvig/app/feature/profile/ui/myhome/ChangeHomeInfoDialog.kt
@@ -1,17 +1,13 @@
 package com.hedvig.app.feature.profile.ui.myhome
 
-import android.content.Intent
 import android.os.Build
 import android.os.Bundle
 import android.support.v4.app.DialogFragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.navigation.findNavController
 import com.hedvig.app.R
-import com.hedvig.app.feature.chat.ChatActivity
 import com.hedvig.app.feature.profile.ui.ProfileViewModel
-import com.hedvig.app.util.extensions.proxyNavigate
 import com.hedvig.app.util.extensions.startClosableChat
 import com.hedvig.app.util.whenApiVersion
 import kotlinx.android.synthetic.main.dialog_change_home_info.*
