+++ android/app/src/main/java/com/hedvig/app/feature/referrals/ReferralsReceiverActivity.kt
@@ -3,10 +3,9 @@ package com.hedvig.app.feature.referrals
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.widget.Toast
 import com.hedvig.app.BaseActivity
 import com.hedvig.app.R
-import com.hedvig.app.feature.chat.ChatActivity
+import com.hedvig.app.feature.chat.NativeChatActivity
 import com.hedvig.app.util.extensions.makeToast
 import com.hedvig.app.util.extensions.observe
 import com.hedvig.app.util.extensions.view.setHapticClickListener
@@ -25,7 +24,7 @@ class ReferralsReceiverActivity : BaseActivity() {
         setContentView(R.layout.referrals_receiver_activity)
 
         referralViewModel.apply {
-            redeemCodeStatus.observe(this@ReferralsReceiverActivity) { data ->
+            redeemCodeStatus.observe(lifecycleOwner = this@ReferralsReceiverActivity) { data ->
                 if (data != null) {
                     startChat()
                 } else {
@@ -53,7 +52,7 @@ class ReferralsReceiverActivity : BaseActivity() {
     }
 
     private fun startChat() {
-        val intent = Intent(this, ChatActivity::class.java)
+        val intent = Intent(this, NativeChatActivity::class.java)
         intent.putExtra("intent", "onboarding")
         intent.putExtra("show_restart", true)
         startActivity(intent)
