--- android/app/src/main/java/com/hedvig/app/react/data/MonthlyDiscount.kt
@@ -1,6 +1,0 @@
-package com.hedvig.app.react.data
-
-data class MonthlyDiscount(
-    val __typename: String,
-    val amount: String
-)
