--- android/app/src/main/java/com/hedvig/app/react/data/MonthlyGross.kt
@@ -1,6 +1,0 @@
-package com.hedvig.app.react.data
-
-data class MonthlyGross(
-    val __typename: String,
-    val amount: String
-)
