--- android/app/src/main/java/com/hedvig/app/react/data/MonthlyNet.kt
@@ -1,6 +1,0 @@
-package com.hedvig.app.react.data
-
-data class MonthlyNet(
-    val __typename: String,
-    val amount: String
-)
