+++ android/app/src/main/res/layout/chat_input_view.xml
@@ -83,7 +83,7 @@
         app:lottie_rawRes="@raw/hedvig_typing_animation_2"
         tools:visibility="visible" />
 
-    <com.hedvig.app.feature.chat.native.AudioRecorderView
+    <com.hedvig.app.feature.chat.AudioRecorderView
         android:id="@+id/audioRecorder"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
